requirements.txt
dask==2024.1.0
easyDataverse==0.3.7
easyDataverse==0.3.7
h5py==3.10.0
importlib_metadata==7.0.1
matplotlib==3.8.2
numpy==1.26.3
pandas==2.2.0
plotly==5.18.0
pytest==7.4.4
PyYAML==6.0.1
PyYAML==6.0.1
scipy==1.12.0
seaborn==0.13.1
setuptools==68.2.2
tqdm==4.65.0
